Hooking Main Field Table
