+++ src/webparts/v2NpmBanner/components/PropPaneCols/components/MainFieldTable.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { useState, useEffect } from 'react';
 import { Panel, PanelType } from 'office-ui-fabric-react/lib/Panel';
 import ReactJson from 'react-json-view';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -16,8 +17,27 @@ import { Icon, } from 'office-ui-fabric-react/lib/Icon';
 
 import { IMinField } from "./IPropPaneColsProps";
 
-export function buildMainFieldTable( filtered: IMinField[], designMode: boolean, listFields: IMinField[], searchProp: string, searchText: string, onSelectItem: any, onTypeClick: any, showFieldPanel: any ) : JSX.Element {
+export interface IMainFieldTableHookProps {
+  // selected: IMinField[];
+  // onKeeperClick: any;
+  // onDirectionClick: any;
 
+  filtered: IMinField[];
+  designMode: boolean;
+  listFields: IMinField[];
+  searchProp: string;
+  searchText: string;
+  onSelectItem: any;
+  onTypeClick: any;
+  showFieldPanel: any;
+
+  // showFieldPanel: any;
+}
+
+// export function buildMainFieldTable( filtered: IMinField[], designMode: boolean, listFields: IMinField[], searchProp: string, searchText: string, onSelectItem: any, onTypeClick: any, showFieldPanel: any ) : JSX.Element {
+const MainFieldTableHook: React.FC<IMainFieldTableHookProps> = ( props ) => {
+
+  const { filtered, designMode, listFields, searchProp, searchText, onSelectItem, onTypeClick, showFieldPanel } = props;
   let heading: string = '';
 
   if ( listFields.length > 0 ) {
@@ -84,47 +104,7 @@ export function buildMainFieldTable( filtered: IMinField[], designMode: boolean,
         { tableRows }
       </table>;
 
-  return MainFieldTable;
+  return  ( MainFieldTable );
 }
 
-export function  getMainSelectedItems ( ev: React.MouseEvent<HTMLElement>, listFields: IMinField[], selected: IMinField[]  ): IMinField []  {
-  const target: any = ev.target;
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { altKey, ctrlKey, shiftKey, type } = ev; // type is like 'click'
-
-  const itemName: string = target.dataset.fieldname;
-  let thisSelected : IMinField = null;
-
-  listFields.map( field => {  //Find selected item
-    if ( field.InternalName === itemName ) { 
-      field.isSelected = field.isSelected === true ? false : true;
-      field.isKeeper = true;
-      thisSelected = field;
-    }
-  });
-
-  let selectedIdx : number = -1;
-  selected.map( ( pick: IMinField, idx : number ) => {
-    if ( pick.InternalName === thisSelected.InternalName ) selectedIdx = idx;
-  });
-
-  let newSelected: IMinField [] = [];
-
-  if ( selectedIdx === -1 ) {  //Add to selected list
-    
-    if ( ctrlKey === true ) {
-      newSelected = [ ...[ thisSelected ], ...selected ];
-    } else {
-      newSelected = [ ...selected, ...[ thisSelected ] ];
-    }
-
-  } else { //Remove from selected list
-    newSelected = selected.filter( (field) => { return field.InternalName !== thisSelected.InternalName } )
-  }
-
-  console.log('_onSelectItem:', itemName, target, newSelected );
-
-  return newSelected;
-}
-
+export default MainFieldTableHook;
